chore(cubesql): Reorganize new test location
